// packages/workflow/src/models/datasets.ts
export enum DataSourceType {
  FILE = "upload_file",
  NOTION = "notion_import",
  WEB = "website_crawl",
}

export type DataSet = {
  id: string
  name: string
  icon: string
  icon_background: string
  description: string
  permission: "only_me" | "all_team_members"
  data_source_type: DataSourceType
  indexing_technique: "high_quality" | "economy"
  created_by: string
  updated_by: string
  updated_at: number
  app_count: number
  document_count: number
  word_count: number
  embedding_model: string
  embedding_model_provider: string
  embedding_available: boolean
  retrieval_model_dict: RetrievalConfig
  retrieval_model: RetrievalConfig
  tags: Tag[]
}